requirements.txt
xgboost
optuna
shap
scikit-learn
pandas
numpy
matplotlib
seaborn
torch
torchvision
torchaudio
mlflow
cookiecutter
mypy
ruff
hydra-core>=1.3
ensorflow==2.15.0
pytest
librosa>=0.10.0
pre-commit
black
ruff
fastapi
uvicorn
scikit-learn==1.6.1
omegaconf
joblib
prometheus-fastapi-instrumentator
starlette
